Changed signature for IntakeManifoldAbsolutePressure. Added fuel system and pid support unit tests

+++ Strados.Obd/ObdParser.cs
@@ -196,7 +196,7 @@ namespace Strados.Obd
             return Convert.ToInt32(data[0], 16) * 3;
         }
 
-        private static double IntakeManifoldAbsolutePressure(string[] data)
+        private static int IntakeManifoldAbsolutePressure(string[] data)
         {
             return Convert.ToInt32(data[0], 16);
         }
